requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.4.1
torchvision
albumentations==0.4.3
pudb==2019.2
imageio==2.9.0
Pillow==10.0.1
imageio-ffmpeg==0.4.2
omegaconf==2.1.1
test-tube>=0.7.5
streamlit>=0.73.1
einops>= 0.4.1
torch-fidelity==0.3.0
transformers==4.30.2
kornia==0.6
webdataset==0.2.5
torchmetrics>= 0.11.0
fire==0.4.0
gradio==3.21.0
diffusers==0.18.2
huggingface-hub==0.16.4
accelerate==0.25.0
datasets==2.14.5
carvekit-colab==4.1.0
rich>=13.3.2
lovely-numpy>=0.2.8
lovely-tensors>=0.1.14
plotly==5.13.1
pytorch-lightning>=1.6.0
wandb
segment_anything==1.0
rembg
opencv-python
ipdb
tqdm
kiui
configargparse
plyfile==1.1
lpips==0.1.4